+++ CHANGELOG.md
@@ -5,6 +5,17 @@ All notable changes from aws-iot-tfg will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
+## [0.5.0] - 29-03-2020
+
+- Added simulator
+- Added Terraform module for IoT Thing
+- Added Lambda in Terraform
+
+## [0.4.0] - 26-03-2020
+
+- Added lambda code
+- Added IoT Topic Rule
+
 ## [0.3.0] - 26-03-2020
 
 - Added role and policy for Lambda